backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
scikit-learn
pandas
numpy
requests
beautifulsoup4
torch
torchvision
pillow
